refactor:  reduce App component complexity from 22 to <15
