+++ src/components/tabs/SettingsTab.tsx
@@ -1,0 +1,84 @@
+import React from 'react';
+import { RefreshCw, Cpu } from 'lucide-react';
+import { enable, disable } from "@tauri-apps/plugin-autostart";
+
+interface SettingsTabProps {
+    isAutostartEnabled: boolean;
+    setIsAutostartEnabled: (enabled: boolean) => void;
+    minimizeToTray: boolean;
+    toggleMinimizeToTray: () => void;
+    latestVersion: string;
+    clientVersion: string;
+    addLog: (msg: string) => void;
+}
+
+const SettingsTab: React.FC<SettingsTabProps> = ({
+    isAutostartEnabled, setIsAutostartEnabled,
+    minimizeToTray, toggleMinimizeToTray,
+    latestVersion, clientVersion, addLog
+}) => {
+    return (
+        <div className="tab-content fadeIn">
+            <div className="card">
+                <h3 className="card-title">Technical Settings</h3>
+                <button type="button" className="settings-row" onClick={async () => {
+                    const newState = !isAutostartEnabled;
+                    if (newState) await enable(); else await disable();
+                    setIsAutostartEnabled(newState);
+                    addLog(`Auto-launch ${newState ? 'enabled' : 'disabled'}.`);
+                }}>
+                    <div className="settings-info">
+                        <span className="settings-label">Auto-launch</span>
+                        <p className="settings-desc">Launch the app automatically when your PC starts.</p>
+                    </div>
+                    <label className="switch">
+                        <input type="checkbox" checked={isAutostartEnabled} readOnly />
+                        <span className="slider"></span>
+                    </label>
+                </button>
+
+                <button type="button" className="settings-row" onClick={toggleMinimizeToTray} style={{ marginTop: '10px' }}>
+                    <div className="settings-info">
+                        <span className="settings-label">Minimize to Tray</span>
+                        <p className="settings-desc">Close button will minimize the app to the system tray.</p>
+                    </div>
+                    <label className="switch">
+                        <input type="checkbox" checked={minimizeToTray} readOnly />
+                        <span className="slider"></span>
+                    </label>
+                </button>
+            </div>
+
+            {latestVersion && clientVersion !== latestVersion && (
+                <div className="card update-panel-hero">
+                    <div className="update-content">
+                        <div className="update-intel">
+                            <RefreshCw size={24} className="intel-spinner" />
+                            <div>
+                                <h3 className="update-title-hero">New Enhancement Available</h3>
+                                <p className="update-desc-hero">A fresh build of the toolkit is ready to be installed (<b>v{latestVersion}</b>).</p>
+                            </div>
+                        </div>
+                        <a href="https://github.com/L9Lenny/lol-profile-editor/releases/latest" target="_blank" rel="noreferrer" className="update-action-btn-hero">
+                            UPDATE NOW
+                        </a>
+                    </div>
+                </div>
+            )}
+
+            <div className="card" style={{ marginTop: '20px', background: 'rgba(200, 155, 60, 0.03)' }}>
+                <div style={{ display: 'flex', alignItems: 'center', gap: '15px' }}>
+                    <Cpu size={24} style={{ color: 'var(--hextech-gold)' }} />
+                    <div>
+                        <h4 style={{ margin: 0, color: 'var(--hextech-gold)', fontSize: '0.9rem' }}>Bridge Interface</h4>
+                        <p style={{ margin: '5px 0 0 0', fontSize: '0.75rem', color: 'var(--text-secondary)' }}>
+                            High-performance LCU communication layer via Tauri v2 Core.
+                        </p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default SettingsTab;
